+++ src/main/java/com/dabsquared/gitlabjenkins/gitlab/hook/model/Project.java
@@ -1,3 +1,6 @@
+/* Note to Reviewers :
+ * will be removed.
+ */
 package com.dabsquared.gitlabjenkins.gitlab.hook.model;
 
 import net.karneim.pojobuilder.GeneratePojoBuilder;
@@ -11,7 +14,7 @@ import org.apache.commons.lang.builder.ToStringBuilder;
 @GeneratePojoBuilder(intoPackage = "*.builder.generated", withFactoryMethod = "*")
 public class Project {
 
-    private Integer id;
+    private Long id;
     private String name;
     private String description;
     private String webUrl;
@@ -121,11 +124,11 @@ public class Project {
         this.httpUrl = httpUrl;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
